# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Опора АПК",
    "description": "Навигатор поддержки. Подбор предварительный. Локальный набор — синтетический.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/demo": {
      "post": {
        "summary": "Demo Login",
        "operationId": "demo_login_api_auth_demo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "parameters": [
          {
            "name": "X-App-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Защита запросов приложения. Origin должен соответствовать PUBLIC_ORIGIN."
          }
        ]
      }
    },
    "/api/auth/max": {
      "post": {
        "summary": "Max Login",
        "operationId": "max_login_api_auth_max_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaxInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "parameters": [
          {
            "name": "X-App-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Защита запросов приложения. Origin должен соответствовать PUBLIC_ORIGIN."
          }
        ]
      }
    },
...
    "/api/profile": {
      "get": {
        "summary": "Profile",
        "operationId": "profile_api_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessProfile-Output"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "put": {
        "summary": "Save Profile",
        "operationId": "save_profile_api_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessProfile-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessProfile-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-App-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Защита запросов приложения. Origin должен соответствовать PUBLIC_ORIGIN."
          }
        ]
      }
    },
...
    "/api/max/webhook": {
      "post": {
        "summary": "Webhook",
        "operationId": "webhook_api_max_webhook_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "maxWebhookSecret": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "BusinessProfile-Input": {
        "properties": {
          "registration_region": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Region"
          },
          "activity_region": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Region"
          },
          "legal_form": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ip",
                  "company",
                  "cooperative",
                  "individual"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Form"
          },
          "is_kfh": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Kfh"
          },
          "tax_regime": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "eshn",
                  "usn",
                  "npd",
                  "general"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Regime"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "crops",
                  "livestock",
                  "processing",
                  "mixed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "equipment",
                  "construction",
                  "working_capital",
                  "consultation"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          },
          "expense_stage": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "planned",
                  "incurred"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Expense Stage"
          },
          "years_active": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Active"
          },
          "own_funds": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000000.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,13}|(?=[\\d.]{1,16}0*$)\\d{0,13}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Own Funds"
          },
          "is_sme": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Sme"
          },
          "special_category": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Category"
          },
          "expense_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 2000.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expense Year"
          },
          "family_kfh_members": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Kfh Members"
          },
          "requested_grant": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000000.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,13}|(?=[\\d.]{1,16}0*$)\\d{0,13}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Grant"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BusinessProfile"
      },
      "BusinessProfile-Output": {
        "properties": {
          "registration_region": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Region"
          },
          "activity_region": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Region"
          },
          "legal_form": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ip",
                  "company",
                  "cooperative",
                  "individual"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Form"
          },
          "is_kfh": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Kfh"
          },
          "tax_regime": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "eshn",
                  "usn",
                  "npd",
                  "general"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Regime"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "crops",
                  "livestock",
                  "processing",
                  "mixed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "equipment",
                  "construction",
                  "working_capital",
                  "consultation"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          },
          "expense_stage": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "planned",
                  "incurred"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Expense Stage"
          },
          "years_active": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Active"
          },
          "own_funds": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,13}|(?=[\\d.]{1,16}0*$)\\d{0,13}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Own Funds"
          },
          "is_sme": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Sme"
          },
          "special_category": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Category"
          },
          "expense_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 2000.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expense Year"
          },
          "family_kfh_members": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Kfh Members"
          },
          "requested_grant": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,13}|(?=[\\d.]{1,16}0*$)\\d{0,13}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Grant"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BusinessProfile"
      },
...
      "DemoInput": {
        "properties": {
          "persona": {
            "type": "string",
            "enum": [
              "farmer",
              "editor",
              "second"
            ],
            "title": "Persona",
            "default": "farmer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DemoInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaxInput": {
        "properties": {
          "init_data": {
            "type": "string",
            "maxLength": 16384,
            "minLength": 1,
            "title": "Init Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "init_data"
        ],
        "title": "MaxInput"
      },
...
      "SessionView": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserView"
          },
          "csrf": {
            "type": "string",
            "title": "Csrf"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user",
          "csrf"
        ],
        "title": "SessionView"
      },
...
      "UserView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "demo": {
            "type": "boolean",
            "title": "Demo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "demo"
        ],
        "title": "UserView"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "opora_session",
        "description": "HttpOnly cookie после /api/auth/max; /api/auth/demo только локально"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Значение csrf из входа или /api/auth/me"
      },
      "maxWebhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Max-Bot-Api-Secret"
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Текущий сервер"
    }
  ]
}